+++ mcu-util/Cargo.toml
@@ -16,7 +16,7 @@ clap.workspace = true
 color-eyre.workspace = true
 crc32fast = "1.3.2"
 image = "0.24.8"
-orb-mcu-protocol.path = "../mcu-protocol"
+orb-mcu-interface.path = "../mcu-interface"
 tokio.workspace = true
 tracing-subscriber.workspace = true
 tracing.workspace = true
