+++ mcu-util/src/orb/dfu.rs
@@ -1,6 +1,6 @@
 use color_eyre::eyre::{eyre, Result};
-use orb_mcu_protocol::orb_messages::mcu_main as main_messaging;
-use orb_mcu_protocol::orb_messages::mcu_sec as sec_messaging;
+use orb_mcu_interface::orb_messages::mcu_main as main_messaging;
+use orb_mcu_interface::orb_messages::mcu_sec as sec_messaging;
 use std::cmp::min;
 use std::fs::File;
 use std::io;
