+++ mcu-util/src/orb/security_board.rs
@@ -6,10 +6,10 @@ use std::time::Duration;
 use tokio::time;
 use tracing::{debug, info, warn};
 
-use orb_mcu_protocol::can::canfd::CanRawMessaging;
-use orb_mcu_protocol::can::isotp::{CanIsoTpMessaging, IsoTpNodeIdentifier};
-use orb_mcu_protocol::orb_messages;
-use orb_mcu_protocol::{Device, McuPayload, MessagingInterface};
+use orb_mcu_interface::can::canfd::CanRawMessaging;
+use orb_mcu_interface::can::isotp::{CanIsoTpMessaging, IsoTpNodeIdentifier};
+use orb_mcu_interface::orb_messages;
+use orb_mcu_interface::{Device, McuPayload, MessagingInterface};
 use orb_messages::mcu_sec::battery_status::BatteryState;
 use orb_messages::{mcu_sec as security_messaging, CommonAckError};
